requirements.txt
flask
flask-cors
pyjwt
werkzeug
python-dotenv
pytest